Move code from cpp/ and headers/ into src/engine/
