+++ src/engine/asset.cpp
