+++ src/engine/asset.h
