+++ src/engine/canonical_position.cpp
