+++ src/engine/canonical_position.h
