+++ src/engine/debug_visualize.cpp
