+++ src/engine/gpu_mapped_buffer.cpp
