+++ src/engine/loaders/collada.cpp
