+++ src/engine/loaders/obj.cpp
