+++ src/engine/threadsafe.cpp
