+++ src/engine/triangle.cpp
