+++ src/engine/work_queue.h
@@ -122,6 +122,9 @@ WorkQueueEntry( particle_system *System, untextured_3d_geometry_buffer *Dest, v3
 
 
 
+// TODO(Jesse): This should be able to go into the stdlib at this point.
+// Probably time to move it there.  AFAIK the only thinig that has to move with
+// it is this bonsai_worker_thread_callback thingy.
 link_internal void
 DrainQueue(work_queue* Queue, thread_local_state* Thread, bonsai_worker_thread_callback GameWorkerThreadCallback)
 {
